+++ app/Models/User.php
@@ -157,7 +157,7 @@ class User extends Model implements
     protected $attributes = [
         'external_id' => null,
         'root_admin' => false,
-        'language' => 'en',
+        'language' => 'ja',
         'use_totp' => false,
         'totp_secret' => null,
     ];